project structure changed

+++ Data/Repositories/RepositoryImpl/ProjectRepository.cs
@@ -1,19 +1,20 @@
-﻿using Dapper;
+﻿using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Dapper;
 using Microsoft.Extensions.Configuration;
 using pm.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Runtime.InteropServices;
-using System.Threading.Tasks;
+using Task = System.Threading.Tasks.Task;
 
-namespace project_managment.Services
+namespace project_managment.Services.RepositoryImpl
 {
     public class ProjectRepository : BaseRepository, IProjectRepository
     {
         private const string ProjectMappingString = "id as Id, name as Name, creator_id as CreatorId, description as Description, is_private as IsPrivate, created_at as CreatedAt";
-        private const string TableFieldsString = "name, description, created_at, creator_id, is_private";
-        private const string ObjectFieldsString = "@Name, @Description, @CreatedAt, @CreatorId, @IsPrivate";
+        private const string TableFieldsString = "id, name, description, created_at, creator_id, is_private";
+        private const string ObjectFieldsString = "@Id, @Name, @Description, @CreatedAt, @CreatorId, @IsPrivate";
+        private const string TableFieldsWithoutIdString = "name, description, created_at, creator_id, is_private";
+        private const string ObjectFieldsWithoutIdString = "@Name, @Description, @CreatedAt, @CreatorId, @IsPrivate";
         private const string TableName = "projects";
 
         public ProjectRepository(IConfiguration configuration) : base(configuration)
@@ -38,27 +39,33 @@ namespace project_managment.Services
 
         public async Task<Project> FindById(long id)
         {
-            string sql = $@"SELECT {ProjectMappingString} FROM {TableName} WHERE id = @id";
+            var sql= $@"SELECT {ProjectMappingString} FROM {TableName} WHERE id = @id";
             return await WithConnection(async (connection) => await connection.QueryFirstOrDefaultAsync<Project>(sql, new { id = id }));
         }
 
         public async Task<IEnumerable<Project>> FindNotPrivateProjects()
         {
-            string sql = $@"SELECT {ProjectMappingString} FROM {TableName} WHERE is_private = False";
+            var sql= $@"SELECT {ProjectMappingString} FROM {TableName} WHERE is_private = False";
 
             return await WithConnection(async (connection) => await connection.QueryAsync<Project>(sql));
         }
 
-        public async Task<IEnumerable<Project>> FindProjectsByName(string Name)
+        public async Task<IEnumerable<Project>> FindProjectsByName(string name)
         {
-            string sql = $@"SELECT {ProjectMappingString} FROM {TableName} WHERE name = @Name";
+            var sql= $@"SELECT {ProjectMappingString} FROM {TableName} WHERE name = @Name";
 
-            return await WithConnection<IEnumerable<Project>>(async (connection) => await connection.QueryAsync<Project>(sql, new { Name }));
+            return await WithConnection<IEnumerable<Project>>(async (connection) => await connection.QueryAsync<Project>(sql, new
+                {
+                    Name = name
+                }));
         }
 
         public async  System.Threading.Tasks.Task Remove(Project entity)
         {
-            string sql = $@"DELETE FROM {TableName} WHERE id = @id";
+            if (entity?.Id == null)
+                throw new Exception();
+            
+            var sql= $@"DELETE FROM {TableName} WHERE id = @id";
 
             await WithConnection(async (connection) =>
             {
@@ -68,7 +75,7 @@ namespace project_managment.Services
 
         public async System.Threading.Tasks.Task RemoveById(long id)
         {
-            string sql = $@"DELETE FROM {TableName} WHERE id = @id";
+            var sql= $@"DELETE FROM {TableName} WHERE id = @id";
 
             await WithConnection(async (connection) =>
             {
@@ -78,8 +85,11 @@ namespace project_managment.Services
 
         public async System.Threading.Tasks.Task Save(Project entity)
         {
-            string sql = $@"INSERT INTO {TableName}({TableFieldsString}) VALUES " +
-                         $@"({ObjectFieldsString})";
+            if (entity?.Id == null)
+                throw new Exception();
+                
+            var sql= $@"INSERT INTO {TableName}({TableFieldsWithoutIdString}) VALUES " +
+                         $@"({ObjectFieldsWithoutIdString})";
             await WithConnection(async (connection) =>
             {
                 await connection.ExecuteAsync(sql, entity);
@@ -88,11 +98,26 @@ namespace project_managment.Services
 
         public async System.Threading.Tasks.Task Update(Project entity)
         {
-            string sql = $@"UPDATE {TableName} SET ({TableFieldsString}) = ({ObjectFieldsString}) WHERE id = @Id";
+            if (entity?.Id == null)
+                throw new Exception();
+            
+            var sql= $@"UPDATE {TableName} SET ({TableFieldsString}) = ({ObjectFieldsString}) WHERE id = @Id";
 
             await WithConnection(async (connection) =>
             {
                 await connection.ExecuteAsync(sql, entity);
+            });
+        }
+
+        public async Task LinkUserAndProject(User user, Project project) // подходит ли метод под паттерн репозитория ?
+        {
+            if (user?.Id == null || project?.Id == null)
+                throw new Exception();
+            
+            var sql = $@"INSERT INTO project_user(project_id, user_id) VALUES(@ProjectId, @UserId)";
+            await WithConnection(async (connection) =>
+            {
+                await connection.ExecuteAsync(sql, new {ProjectId = project.Id, UserId = user.Id});
             });
         }
     }
